feat: use Taskfile to manage tasks
