+++ .github/workflows/main.yml
@@ -27,11 +27,10 @@ jobs:
      - name: Run Ruff (linting)
        uses: astral-sh/ruff-action@v3
 
-     - name: Run Alembic migrations
-       run: alembic upgrade head
-
      - name: Run pytest (unit tests)
-       run: uv run pytest
+       env:
+        DB_PATH: test.db
+       run: uv run alembic upgrade head && uv run pytest
 
      - name: Run mypy (type checking)
        run: uv run mypy app/
