+++ alembic/versions/73d2e0f178ae_auto.py
@@ -1,0 +1,43 @@
+"""auto
+
+Revision ID: 73d2e0f178ae
+Revises: 
+Create Date: 2025-06-30 01:02:13.350995
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+import sqlmodel
+
+
+# revision identifiers, used by Alembic.
+revision: str = '73d2e0f178ae'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('csvmetadata',
+    sa.Column('name_stored', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('name_original', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('size_bytes', sa.Integer(), nullable=False),
+    sa.Column('nrows', sa.Integer(), nullable=False),
+    sa.Column('ncols', sa.Integer(), nullable=False),
+    sa.Column('uploaded_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('name_stored')
+    )
+    op.create_index(op.f('ix_csvmetadata_name_stored'), 'csvmetadata', ['name_stored'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_csvmetadata_name_stored'), table_name='csvmetadata')
+    op.drop_table('csvmetadata')
+    # ### end Alembic commands ###
