--- alembic/versions/bd09512f9f99_initial_migration.py
@@ -1,28 +1,0 @@
-"""initial migration
-
-Revision ID: bd09512f9f99
-Revises: 
-Create Date: 2025-06-27 13:25:58.380864
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'bd09512f9f99'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    pass
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    pass
