--- pytest.ini
@@ -1,4 +1,0 @@
-[pytest]
-env =
-    ENV=test
-    DB_PATH=test.db
